+++ widgetbook/lib/main.directories.g.dart
@@ -536,6 +536,11 @@ final directories = <_widgetbook.WidgetbookNode>[
             builder:
                 _widgetbook_workspace_adaptive_cards_use_cases.buildV16Rating,
           ),
+          _widgetbook.WidgetbookUseCase(
+            name: 'Rating input overlay (knob)',
+            builder: _widgetbook_workspace_adaptive_cards_use_cases
+                .buildRatingInputOverlay,
+          ),
         ],
       ),
       _widgetbook.WidgetbookComponent(